refactor(rag): import RagStatus from app.lib.rag in client modules
Client/plugin code should consume RagStatus through the public boundary,
not reach into app.rag.enums. Leaves the facade, RAG-internal modules,
app.models.drive (circular import), and the deferred rag_mcp on app.rag.enums.
