+++ app/core_plugins/slack/rag.py
@@ -22,10 +22,10 @@ from app.core_plugins.slack.models import ResponseType
 from app.core_plugins.slack.synthesis import synthesize_answer
 from app.lib.db import session_scope
 from app.lib.log import get_logger
+from app.lib.rag import RagStatus
 from app.llm.providers import BaseChatProvider
 from app.models.drive import DriveFile
 from app.rag.context_service import RAGContextService, format_chunks_as_context
-from app.rag.enums import RagStatus
 from app.rag.exceptions import (
     DriveFileNotFoundError,
     RAGNotReadyError,
